refactor: ai modules split
